Fix signup and signin
