+++ frontend/components/radialProfileCard.tsx
@@ -43,18 +43,6 @@ const RadialProfileCard = () => {
   const circumference = 2 * Math.PI * radius;
   const offset = circumference - (completionPercentage / 100) * circumference;
 
-  useEffect(() => {
-    const getUserData = async () => {
-      const data = await fetchUserData();
-      setUserData(data);
-      const percentage = calculateProfileCompletion(data);
-      setCompletionPercentage(percentage);
-      updateStatusMessage(percentage);
-      updateCompletionStatus(data);
-    };
-    getUserData();
-  }, []);
-
   const updateStatusMessage = (percentage: number) => {
     let msg = '';
     if (percentage === 100) {
@@ -73,18 +61,32 @@ const RadialProfileCard = () => {
 
   const updateCompletionStatus = (data: any) => {
     setCompletionStatus({
-      name: !!data.basics?.name,
-      label: !!data.basics?.label,
-      image: !!data.basics?.image,
-      email: !!data.basics?.email,
-      phone: !!data.basics?.phone,
-      url: !!data.basics?.url,
-      githubUrl: !!data.projects[0]?.githubUrl,
-      education: !!data.education[0]?.institution,
-      certificates: !!data.certificates[0]?.name,
-      skills: data.projects[0]?.techStack.length > 2,
+      name: !!data?.basics?.name,
+      label: !!data?.basics?.label,
+      image: !!data?.basics?.image,
+      email: !!data?.basics?.email,
+      phone: !!data?.basics?.phone,
+      url: !!data?.basics?.url,
+      githubUrl: !!data?.projects[0]?.githubUrl,
+      education: !!data?.education[0]?.institution,
+      certificates: !!data?.certificates[0]?.name,
+      skills: data?.projects[0]?.techStack.length > 2,
     });
   };
+
+  useEffect(() => {
+    const getUserData = async () => {
+      const data = await fetchUserData();
+      setUserData(data);
+      const percentage = calculateProfileCompletion(data);
+      setCompletionPercentage(percentage);
+      updateStatusMessage(percentage);
+      updateCompletionStatus(data);
+    };
+    getUserData();
+  }, []);
+
+  console.log(userData)
 
   return (
     <div className="flex flex-col w-2/12 mx-3 mr-3">
