+++ frontend/types/next-auth.d.ts
@@ -1,0 +1,22 @@
+import 'next-auth'
+
+declare module "next-auth" {
+    interface Session {
+        user: {
+            id?: string;
+            email?: string;
+        } & DefaultSession['user']
+    }
+
+    interface User {
+        id?: string;
+        email?: string;
+    }
+}
+
+declare module "next-auth/jwt" {
+    interface JWT {
+        id?: string;
+        email?: string;
+    }
+}
